feat(recruitment): enhance job requisition factory to create stages a… (#75)
This pull request updates the job requisition testing and factory logic
to better reflect the requirements for available jobs—specifically,
ensuring that only jobs with stages and the correct status are
considered available. The changes improve the reliability and accuracy
of tests and test data.

**Testing improvements:**

* The `SearchJobsTest` now explicitly creates non-available jobs with
the `Draft` status, making the test more robust and clear about which
jobs should not be rendered as available.
[[1]](diffhunk://#diff-9216231eed59c52a1856544428702cdd81051a8ee8d25ec37b051d631e7df54cR6)
[[2]](diffhunk://#diff-9216231eed59c52a1856544428702cdd81051a8ee8d25ec37b051d631e7df54cR15-R21)

**Factory enhancements:**

* The `JobRequisitionFactory`'s `available` state now automatically
creates between 2 and 4 related `Stage` models for each available job
requisition, ensuring that test data better matches the application's
expectations for available jobs.
[[1]](diffhunk://#diff-366e099df1ca5cd567015efe61b25e97abbcfeb170a429cec5f4a247a64d0392R15)
[[2]](diffhunk://#diff-366e099df1ca5cd567015efe61b25e97abbcfeb170a429cec5f4a247a64d0392R75-R81)…nd
update test for job availability

+++ app-modules/panel-app/src/Livewire/MyProfile/CandidateSkills.php
@@ -58,7 +58,7 @@ class CandidateSkills extends MyProfileComponent
                             ->label(__('panel-app::pages/settings.skills.fields.years_of_experience'))
                             ->prefixIcon('heroicon-o-calendar-days')
                             ->suffix(__('panel-app::pages/settings.skills.placeholders.years_suffix'))
-                            ->numeric()
+                            ->integer()
                             ->minValue(0)
                             ->maxValue(50)
                             ->placeholder('0')
